frontend/requirements.txt
streamlit==1.26.0
requests~=2.31.0
pyyaml==6.0.1
matplotlib==3.5.0
Pillow~=9.1.0
scikit-image~=0.21.0